app: vrbox: mqtt: May fix crash on admin_fetch command
The original mqtt callback function is running in one of the system
thread. The document suggest that the callback function should be
processing data as fast as it can.

We assume that the crash(Dump message: Illegal load of EXC_RETURN) on
admin_fetch function may caused by this issue, run time too long on
system thread. This fix never tested, need a little assistant of upper
computer to test it.

Note: The network reconnection still not soled. This bug may leads crash
on HTTP when the board needs to download new firmware via HTTP.

Signed-off-by: Ding Tao <miyatsu@qq.com>