app: vrbox: axle: Fix no effect function axle_rotate_to()
After the axle is initialized, the function axle_rotate_to() did not
work as we expected. This patch fixed this bug and fully tested.

Note: One bug is still not been fixed: Once the axle position is at
position 3, the irq may trigged when axle move out of position 3 because
the position 3 some how related to position 1 and 2.

Fix plan: Move out of the certain position before enable the position
irq in case of trigger the irq at falling edge from different pin.

Signed-off-by: Ding Tao <miyatsu@qq.com>