app: vrbox: door: Fix possible memory consistency problem
When the infrared irq enabled, the irq handler will write true into one
of the infrared detected flag. This flag array will be checked by the
main handle thread, which may cause a memory consistency problem when
access it twice.

Fix this problem by replace the polling method with semaphore to sync
two threads main handle thread and irq handler. To achive this, we have
to change a new API to door open/close function.

All basic functions are tested include the infrared detector. It may not
conver all possible situations, but it still better than previous
version of this code.

Note: The admin_open() and admin_close() are never tested since this
commit. The door open and close pin triggered when its got long trigger
time, it MAY not be triggered if the trigger time is too short.

Signed-off-by: Ding Tao <miyatsu@qq.com>