app: vrbox: mqtt: Fix possible missing incomming cmd
In order to not block the system, we use a thread to handle the in
comming comand and the network callback return immediately. At the
callback processing stage, we did not store the in comming command. This
may cause lose of message.

Fix this by using a fifo as command queue, repalce handle thread with
dispatch thread polling the fifo.

In test stage we find the close command MAY missing during the message
transmissing, fix this by change the MQTT QoS to QoS 2.

After one round test, it is successful checked.

Signed-off-by: Ding Tao <miyatsu@qq.com>