app: vrbox: dfu: Change http API to net_app
The old HTTP api is deprecated, we change the HTTP api to newest http
network stack.

The request processing is successfuly tested on STM32F407 SoC, but the
response callback funtion is still old. We need be sure that the
response function is also change to the newest HTTP api.

Signed-off-by: Ding Tao <miyatsu@qq.com>